+++ src/app/core/services/event-emitter/event-emitter.service.ts
@@ -1,0 +1,16 @@
+import { Injectable, EventEmitter } from "@angular/core";
+import { Subscription } from "rxjs";
+
+@Injectable({
+  providedIn: "root",
+})
+export class EventEmitterService {
+  invokeSaveFile = new EventEmitter();
+  subsVar: Subscription;
+
+  constructor() {}
+
+  onSaveFileTrigger() {
+    this.invokeSaveFile.emit();
+  }
+}